Change in model: use Long and Double instead of BigInteger and BigDecimal (problems could appear with rounding) + handle frequency table on dates
git-svn-id: http://talendforge.org/svn/tis_shared/trunk@14418 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.cwm.management/src/test/java/org/talend/dq/indicators/LoadSerialData.java
@@ -13,7 +13,6 @@
 package org.talend.dq.indicators;
 
 import java.io.File;
-import java.math.BigInteger;
 
 import org.eclipse.emf.common.util.EList;
 import org.eclipse.emf.common.util.URI;
@@ -58,7 +57,7 @@ public class LoadSerialData {
         for (EObject object : contents) {
             FrequencyIndicator freqI = mySwitch.doSwitch(object);
             if (freqI != null) {
-                BigInteger uniqueValueCount = freqI.getUniqueValueCount();
+                Long uniqueValueCount = freqI.getUniqueValueCount();
                 System.out.println("nb unique values = " + uniqueValueCount);
                 EList<Object> uniqueValues = freqI.getUniqueValues();
                 for (Object data : uniqueValues) {
